% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midasflow.R
\name{mefa}
\alias{mefa}
\title{Run the Memory-Efficient Flow Accumulation (MEFA) algorithm}
\usage{
mefa(
  dir_path,
  accum_path,
  dir_opts = NULL,
  encoding = NULL,
  use_lessmem = TRUE,
  compress_output = FALSE,
  num_threads = 0
)
}
\arguments{
\item{dir_path}{Character. Input flow direction raster (e.g., gpkg:file.gpkg:layer).}

\item{accum_path}{Character. Output flow accumulation GeoTiff.}

\item{dir_opts}{Character, optional. Comma-separated list of GDAL options for dir_path.}

\item{encoding}{Character, optional. Input flow direction encoding.

power2 (default): \eqn{2^{0,\cdots,7}} CW from E (e.g., r.terraflow, ArcGIS)

taudem: 1-8 (E-SE CCW) (e.g., d8flowdir)

45degree: 1-8 (NE-E CCW) (e.g., r.watershed)

degree: (0,360] (E-E CCW)

E,SE,S,SW,W,NW,N,NE: custom (e.g., 1,8,7,6,5,4,3,2 for taudem)}

\item{use_lessmem}{Logical, optional. Use less memory.}

\item{compress_output}{Logical, optional. Compress output GeoTIFF.}

\item{num_threads}{Integer, optional. Number of threads (default OMP_NUM_THREADS).}
}
\value{
Integer status code (0 for success).
}
\description{
Run the Memory-Efficient Flow Accumulation (MEFA) algorithm
}
\references{
Cho, H. (2023). \emph{Memory-Efficient Flow Accumulation Using a Look-Around Approach and Its OpenMP Parallelization}. Environmental Modelling & Software 167, 105771. \url{https://doi.org/10.1016/j.envsoft.2023.105771}.
}
