% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midasflow.R
\name{melfp}
\alias{melfp}
\title{Run the Memory-Efficient Longest Flow Path (MELFP) algorithm}
\usage{
melfp(
  dir_path,
  outlets_path,
  id_col,
  output_path,
  dir_opts = NULL,
  encoding = NULL,
  outlets_layer = NULL,
  outlets_opts = NULL,
  oid_col = "lfp_id",
  lfp_name = "lfp",
  heads_name = NULL,
  coors_path = NULL,
  find_full = FALSE,
  use_lessmem = 2,
  save_outlets = FALSE,
  num_threads = 0,
  tracing_stack_size = 0
)
}
\arguments{
\item{dir_path}{Character. Input flow direction raster (e.g., gpkg:file.gpkg:layer).}

\item{outlets_path}{Character. Input outlets vector.}

\item{id_col}{Character. Input column for outlet IDs.}

\item{output_path}{Character. Output watersheds GeoTIFF or output text file with save_outlets.}

\item{dir_opts}{Character, optional. Comma-separated list of GDAL options for dir_path.}

\item{encoding}{Character, optional. Input flow direction encoding.

power2 (default): \eqn{2^{0,\cdots,7}} CW from E (e.g., r.terraflow, ArcGIS)

taudem: 1-8 (E-SE CCW) (e.g., d8flowdir)

45degree: 1-8 (NE-E CCW) (e.g., r.watershed)

degree: (0,360] (E-E CCW)

E,SE,S,SW,W,NW,N,NE: custom (e.g., 1,8,7,6,5,4,3,2 for taudem)}

\item{outlets_layer}{Character, optional. Layer name of input outlets vector, if necessary (e.g., gpkg).}

\item{outlets_opts}{Character, optional. Comma-separated list of GDAL options for outlets.}

\item{oid_col}{Character, optional. Output column for outlet IDs.}

\item{lfp_name}{Character, optional. Layer name for output longest flow paths.}

\item{heads_name}{Character, optional. Layer name for output longest flow path heads.}

\item{coors_path}{Character, optional. Output longest flow path head coordinates CSV.}

\item{find_full}{Logical, optional. Find full longest flow paths.}

\item{use_lessmem}{Integer, optional. Use less memory.}

\item{save_outlets}{Logical, optional. Write outlet rows and columns, and exit.}

\item{num_threads}{Integer, optional. Number of threads (default OMP_NUM_THREADS).}

\item{tracing_stack_size}{Integer, optional. Tracing stack size (default 3072, 0 for guessing).}
}
\value{
Integer status code (0 for success).
}
\description{
Run the Memory-Efficient Longest Flow Path (MELFP) algorithm
}
\references{
Cho, H. (2025). \emph{Loop Then Task: Hybridizing OpenMP Parallelism to Improve Load Balancing and Memory Efficiency in Continental-Scale Longest Flow Path Computation}. Environmental Modelling & Software 193, 106630. \url{https://doi.org/10.1016/j.envsoft.2025.106630}.
}
