% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midasflow.R
\name{meshed}
\alias{meshed}
\title{Run the Memory-Efficient Watershed Delineation (MESHED) algorithm}
\usage{
meshed(
  dir_path,
  outlets_path,
  id_col,
  output_path,
  dir_opts = NULL,
  encoding = NULL,
  outlets_layer = NULL,
  outlets_opts = NULL,
  hier_path = NULL,
  use_lessmem = FALSE,
  compress_output = FALSE,
  save_outlets = FALSE,
  num_threads = 0,
  tracing_stack_size = 0
)
}
\arguments{
\item{dir_path}{Character. Input flow direction raster (e.g., gpkg:file.gpkg:layer).}

\item{outlets_path}{Character. Input outlets vector.}

\item{id_col}{Character. Input column for outlet IDs.}

\item{output_path}{Character. Output watersheds GeoTIFF or output text file with save_outlets.}

\item{dir_opts}{Character, optional. Comma-separated list of GDAL options for dir_path.}

\item{encoding}{Character, optional. Input flow direction encoding.

power2 (default): \eqn{2^{0,\cdots,7}} CW from E (e.g., r.terraflow, ArcGIS)

taudem: 1-8 (E-SE CCW) (e.g., d8flowdir)

45degree: 1-8 (NE-E CCW) (e.g., r.watershed)

degree: (0,360] (E-E CCW)

E,SE,S,SW,W,NW,N,NE: custom (e.g., 1,8,7,6,5,4,3,2 for taudem)}

\item{outlets_layer}{Character, optional. Layer name of input outlets vector, if necessary (e.g., gpkg).}

\item{outlets_opts}{Character, optional. Comma-separated list of GDAL options for outlets.}

\item{hier_path}{Character, optional. Output subwatershed hierarchy CSV.}

\item{use_lessmem}{Logical, optional. Use less memory.}

\item{compress_output}{Logical, optional. Compress output GeoTIFF.}

\item{save_outlets}{Logical, optional. Write outlet rows and columns, and exit.}

\item{num_threads}{Integer, optional. Number of threads (default OMP_NUM_THREADS).}

\item{tracing_stack_size}{Integer, optional. Tracing stack size (default 3072, 0 for guessing).}
}
\value{
Integer status code (0 for success).
}
\description{
Run the Memory-Efficient Watershed Delineation (MESHED) algorithm
}
\references{
Cho, H. (2025). \emph{Avoid Backtracking and Burn Your Inputs: CONUS-Scale Watershed Delineation Using OpenMP}. Environmental Modelling & Software 183, 106244. \url{https://doi.org/10.1016/j.envsoft.2024.106244}.
}
