% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midasflow.R
\name{meufl}
\alias{meufl}
\title{Run the Memory-Efficient Upstream Flow Length (MEUFL) algorithm}
\usage{
meufl(
  dir_path,
  flen_path,
  dir_opts = NULL,
  encoding = NULL,
  from_one = FALSE,
  use_lessmem = 2,
  compress_output = FALSE,
  num_threads = 0
)
}
\arguments{
\item{dir_path}{Character. Input flow direction raster (e.g., gpkg:file.gpkg:layer).}

\item{flen_path}{Character. Output upstream flow length GeoTiff.}

\item{dir_opts}{Character, optional. Comma-separated list of GDAL options for dir_path.}

\item{encoding}{Character, optional. Input flow direction encoding.

power2 (default): \eqn{2^{0,\cdots,7}} CW from E (e.g., r.terraflow, ArcGIS)

taudem: 1-8 (E-SE CCW) (e.g., d8flowdir)

45degree: 1-8 (NE-E CCW) (e.g., r.watershed)

degree: (0,360] (E-E CCW)

E,SE,S,SW,W,NW,N,NE: custom (e.g., 1,8,7,6,5,4,3,2 for taudem)}

\item{from_one}{Logical, optional. Count from 1.}

\item{use_lessmem}{Integer, optional. 2 for least memory, 1 for less memory, and 0 for more memory.}

\item{compress_output}{Logical, optional. Compress output GeoTIFF.}

\item{num_threads}{Integer, optional. Number of threads (default OMP_NUM_THREADS).}
}
\value{
Integer status code (0 for success).
}
\description{
Run the Memory-Efficient Upstream Flow Length (MEUFL) algorithm
}
\references{
Cho, H. (2026). \emph{Flow in Float: Memory-Efficient Upstream Flow Length Parallel Computation Using an IEEE-754-Based Union Encoding}. Environmental Modelling & Software 204, 107045. \url{https://doi.org/10.1016/j.envsoft.2026.107045}.
}
